+++ spotifywebapipython/zeroconfapi/helpers.py
@@ -1,4 +1,5 @@
 from base64 import b64decode, b64encode
+import binascii
 
 
 def write_int(i: int, out: bytearray):
@@ -14,17 +15,43 @@ def write_bytes(b: bytes, out: bytearray):
     out.extend(b)
 
 
-def byte_list_to_int(byte_list):
+def byte_list_to_int(byte_list) -> int:
+    """
+    Converts a list of bytes to a big integer.
+    """
     return int.from_bytes(bytes(byte_list), 'big')
 
 
-def int_to_bytes(i: int):
-    return i.to_bytes((i.bit_length() + 7) // 8, byteorder='big')
+def int_to_bytes(value: int) -> bytes:
+    """
+    Converts a big integer to bytes.
+    """
+    return value.to_bytes((value.bit_length() + 7) // 8, byteorder='big') or b'\0'
 
 
-def int_to_b64str(i: int):
-    return b64encode(int_to_bytes(i)).decode('ascii')
+def int_to_b64str(value: int) -> str:
+    """
+    Converts a big integer to a base64-encoded string value.
+    """
+    return b64encode(int_to_bytes(value)).decode('ascii')
 
 
-def b64_to_int(b64: str):
-    return int.from_bytes(b64decode(b64), 'big')
+def b64_to_int(value: str):
+    """
+    Converts a base64-encoded string or byte array to a big integer.
+    """
+    return int.from_bytes(b64decode(value), 'big')
+
+
+def string_to_int(value):
+    """
+    Converts a string to a big integer.
+    """
+    return int(binascii.hexlify(value), 16)
+
+
+def int_of_string(value):
+    """
+    Converts a string to a big integer.
+    """
+    return int(binascii.hexlify(value), 16)